Run tests with Anvil.
Update the test framework and tests to use Xaya X and Anvil to run
all tests in an EVM environment.
