+++ ejabberd/Makefile.am
@@ -1,5 +1,5 @@
 AM_TESTS_ENVIRONMENT = \
-  PYTHONPATH=$(PYTHONPATH):$(srcdir):$(top_srcdir)/gametest
+  PYTHONPATH=$(srcdir):$(top_srcdir)/gametest
 
 TEST_FILES = \
   tests/integration.py \
