+++ requirements.txt
@@ -1,2 +1,2 @@
 jsonrpclib-pelix
-xayagametest
+xayax
